# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for a multi-tenant (SaaS) ophthalmology clinic\nmanagement system: patients, visits, billing/POS, inventory,\nlab jobs, reports, and optical calculators.\n\nEvery route under /api (other than /api/auth/*, /api/signup,\nand the /api/auth/password-reset|email endpoints) requires a\nBearer access token, and most additionally require the\ncaller's clinic to have an active subscription.",
        "title": "Opthalmic Management System API",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/api",
    "paths": {
...
        "/calculators/near-add": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculators"
                ],
                "summary": "Compute near-add power",
                "parameters": [
                    {
                        "description": "Distance sphere + add power",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.NearAddReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ScalarResultRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/calculators/spherical-equivalent": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculators"
                ],
                "summary": "Compute spherical equivalent",
                "parameters": [
                    {
                        "description": "Sphere/cylinder",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SphericalEquivalentReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ScalarResultRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/calculators/telescope-fov": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculators"
                ],
                "summary": "Compute apparent field of view for a telescope/loupe",
                "parameters": [
                    {
                        "description": "True FOV (degrees) + magnification",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TelescopeFOVReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ScalarResultRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/calculators/transposition": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculators"
                ],
                "summary": "Transpose a cylinder between plus/minus cylinder notation",
                "parameters": [
                    {
                        "description": "Sphere/cylinder/axis",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TranspositionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TranspositionRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/calculators/vertex-distance": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculators"
                ],
                "summary": "Adjust power for a change in vertex distance",
                "parameters": [
                    {
                        "description": "Power + from/to vertex distance (mm)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.VertexDistanceReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ScalarResultRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.NearAddReq": {
            "type": "object",
            "properties": {
                "add_power": {
                    "type": "number"
                },
                "distance_sphere": {
                    "type": "number"
                }
            }
        },
...
        "dto.ScalarResultRes": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "number"
                }
            }
        },
        "dto.SphericalEquivalentReq": {
            "type": "object",
            "properties": {
                "cylinder": {
                    "type": "number"
                },
                "sphere": {
                    "type": "number"
                }
            }
        },
...
        "dto.TelescopeFOVReq": {
            "type": "object",
            "properties": {
                "magnification": {
                    "type": "number"
                },
                "true_fov_degrees": {
                    "type": "number"
                }
            }
        },
        "dto.TranspositionReq": {
            "type": "object",
            "properties": {
                "axis": {
                    "type": "integer",
                    "maximum": 180,
                    "minimum": 0
                },
                "cylinder": {
                    "type": "number"
                },
                "sphere": {
                    "type": "number"
                }
            }
        },
        "dto.TranspositionRes": {
            "type": "object",
            "properties": {
                "axis": {
                    "type": "integer"
                },
                "cylinder": {
                    "type": "number"
                },
                "sphere": {
                    "type": "number"
                }
            }
        },
...
        "dto.VertexDistanceReq": {
            "type": "object",
            "properties": {
                "from_distance_mm": {
                    "type": "number"
                },
                "power": {
                    "type": "number"
                },
                "to_distance_mm": {
                    "type": "number"
                }
            }
        },
...
        "http.response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the access token, e.g. \"Bearer eyJhbGci...\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}